switched to float velocity for variety of poo angles and removed all warnings (for explicit conversion)

+++ Engine/Goal.h
@@ -13,14 +13,14 @@ public:
 	{}
 	void Draw( Graphics& gfx ) const
 	{
-		gfx.DrawRectDim( x,y,dimension,dimension,c );
+		gfx.DrawRectDim( int(x),int(y),int(dimension),int(dimension),c );
 	}
 	bool TestCollision( const Dude& dude ) const
 	{
-		const int duderight = dude.GetX() + dude.GetWidth();
-		const int dudebottom = dude.GetY() + dude.GetHeight();
-		const int pooright = x + dimension;
-		const int poobottom = y + dimension;
+		const float duderight = dude.GetX() + dude.GetWidth();
+		const float dudebottom = dude.GetY() + dude.GetHeight();
+		const float pooright = x + dimension;
+		const float poobottom = y + dimension;
 
 		return
 			duderight >= x &&
@@ -59,7 +59,7 @@ public:
 		}
 	}
 private:
-	static constexpr int dimension = 20;
+	static constexpr float dimension = 20.0f;
 	Color c = { 127,0,0 };
 	bool colorIncreasing = true;
 	int x;
